Separate out XIOS domain tests
